apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend with structlog and pydantic"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn==0.29.0",
    "structlog==24.1.0",
    "python-dotenv==1.0.1",
    "httpx==0.27.0",
    "pydantic-settings>=2.9.1",
]
